requirements.txt
apache-airflow-providers-cncf-kubernetes==10.1.0
apache-airflow-providers-docker==4.0.0